Improve a comment on Data/Downloder.fs
The null -> -1 is just for keep more easy to handle the static types
on parsing JSON on frontend. Because, at least, using JsonProvider this
field should return a structure called NumbersOrDateTime, which will
have one array called .Numbers and a field called  .DateTime.

But if some element is null, this element just doesn't appear!
And I don't know which column (currency) was null. This is bad.
Using -1 instead null, I keep the element on the list so, after
parsing, I can just filter -1 to be points to not rendering.